d_a_npc_aj1: chk_areaIN s32->u8 return (retail callers clrlwi.24) — FARwai 98.89->99.84; chk_areaIN stays 95.28 (5 accept-class anchor/scheduling); fresh .o postdates

+++ include/d/actor/d_a_npc_aj1.h
@@ -90,7 +90,7 @@ public:
     void setSmoke();
     int set_action(int (daNpc_Aj1_c::*)(void*), void*);
     void setStt(signed char);
-    s32 chk_areaIN(float, short, cXyz);
+    u8 chk_areaIN(float, short, cXyz);
     s32 FARwai();
     s32 call_1();
     s32 wait_1();
